Added notebook

+++ src/utils.py
@@ -4,6 +4,6 @@ def get_concreteness_for_word(words, df):
 	"""Look up concreteness of each word in dataframe"""
 	cnc = []
 	for word in words:
-		cnc.append(int(df[df['word'] == word]['Concreteness']))
+		cnc.append(float(df[df['word'] == word]['Concreteness']))
 	return cnc
